+++ src/Resources/contao/languages/de/tl_fernschach_turniere.php
@@ -65,6 +65,7 @@ $GLOBALS['TL_LANG']['tl_fernschach_turniere']['new'] = array('Neuer Eintrag', 'N
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['editHeader'] = array('Eintrag bearbeiten', 'Kategorie bzw. Turnier %s bearbeiten');
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['editBewerbungen'] = array('Bewerbungen bearbeiten', 'Bewerbungen des Turniers %s bearbeiten');
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['editMeldungen'] = array('Anmeldungen bearbeiten', 'Anmeldungen des Turniers %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere']['editMannschaften'] = array('Mannschaften bearbeiten', 'Gemeldete Mannschaften des Turniers %s bearbeiten');
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['editTeilnehmer'] = array('Teilnehmer bearbeiten', 'Teilnehmer/Spieler des Turniers %s bearbeiten');
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['copy'] = array('Eintrag kopieren', 'Kategorie bzw. Turnier %s kopieren');
 $GLOBALS['TL_LANG']['tl_fernschach_turniere']['delete'] = array('Eintrag löschen', 'Kategorie bzw. Turnier %s löschen');