+++ src/Resources/contao/languages/de/tl_fernschach_turniere_mannschaften.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * Fernschach-Verwaltung für Contao Open Source CMS
+ *
+ * @author    Frank Hoppe
+ * @license   LGPL-3.0-or-later
+ */
+
+// Legenden
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['mannschaft_legend'] = 'Mannschaft';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['fuehrer_legend'] = 'Mannschaftsleiter';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['meldung_legend'] = 'Meldung';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['publish_legend'] = 'Aktivieren';
+
+// Felder
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['vereinsname'] = array('Verein', 'Name des Vereins oder der Spielgemeinschaft, für die die Mannschaft antritt.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['vereinsnameAlt'] = array('Alter Vereinsname', 'Nur ausgefüllt, wenn der Verein umbenannt wurde.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['mannschaftsname'] = array('Mannschaft', 'Bezeichnung der Mannschaft, bei mehreren Teams mit römischer Ziffer, etwa „Musterstadt II".');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['spielerId'] = array('Mannschaftsleiter', 'Der Spieler, der die Mannschaft gemeldet hat. Ihm wird das Nenngeld belastet.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['memberId'] = array('BdF-Mitgliedsnummer', 'Mitgliedsnummer des Mannschaftsleiters zum Zeitpunkt der Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['email'] = array('E-Mail', 'E-Mail-Adresse des Mannschaftsleiters zum Zeitpunkt der Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['meldungDatum'] = array('Meldedatum', 'Zeitpunkt, zu dem die Mannschaft gemeldet wurde.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['nenngeld'] = array('Nenngeld', 'Das Nenngeld, das dem Mannschaftsleiter belastet wurde.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['bemerkungen'] = array('Bemerkungen', 'Freie Anmerkungen des Mannschaftsleiters zur Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['published'] = array('Aktiv', 'Aktivieren oder deaktivieren Sie hier die Mannschaftsmeldung.');
+
+// Schaltflächen
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['edit'] = array('Aufstellung bearbeiten', 'Aufstellung der Mannschaft %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['editheader'] = array('Mannschaft bearbeiten', 'Angaben der Mannschaft %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['delete'] = array('Mannschaft löschen', 'Mannschaft %s löschen');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['deleteConfirm'] = 'Die Aufstellung und die zugehörenden Nenngeldsätze werden dabei mitgelöscht.';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['toggle'] = array('Mannschaft veröffentlichen', 'Mannschaft %s veröffentlichen oder verbergen');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften']['show'] = array('Details anzeigen', 'Details der Mannschaft %s anzeigen');