+++ src/Resources/contao/languages/de/tl_fernschach_turniere_mannschaften_spieler.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * Fernschach-Verwaltung für Contao Open Source CMS
+ *
+ * @author    Frank Hoppe
+ * @license   LGPL-3.0-or-later
+ */
+
+// Legenden
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['brett_legend'] = 'Brett';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['spieler_legend'] = 'Angaben zum Spieler';
+
+// Felder
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['brett'] = array('Brett', 'Nummer des Bretts in der Aufstellung, beginnend bei 1.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['spielerId'] = array('Spieler', 'Der Spieler aus dem Spielerregister. Beim Speichern werden Name und Nummern von dort übernommen.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['vorname'] = array('Vorname', 'Vorname zum Zeitpunkt der Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['nachname'] = array('Nachname', 'Nachname zum Zeitpunkt der Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['memberId'] = array('BdF-Mitgliedsnummer', 'Mitgliedsnummer zum Zeitpunkt der Meldung.');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['memberInternationalId'] = array('ICCF-ID', 'Internationale Kennung zum Zeitpunkt der Meldung.');
+
+// Schaltflächen
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['edit'] = array('Brett bearbeiten', 'Brett %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['delete'] = array('Brett löschen', 'Brett %s löschen');
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['deleteConfirm'] = 'Der zugehörende Nenngeldsatz über 0 € wird dabei mitgelöscht.';
+$GLOBALS['TL_LANG']['tl_fernschach_turniere_mannschaften_spieler']['show'] = array('Details anzeigen', 'Details des Bretts %s anzeigen');